+++ apps/qna/serializers/question/response.py
@@ -2,10 +2,17 @@ from typing import Any
 
 from rest_framework import serializers
 
-from apps.qna.models import Question, QuestionCategory
-from apps.qna.serializers.question_base import QnAVersionedValidationMixin
+from apps.qna.models import Question
+from apps.qna.serializers.base import QnAVersionedValidationMixin
+from apps.qna.serializers.question.common import (
+    QuestionAuthorSerializer,
+    QuestionCategoryListSerializer,
+)
 from apps.qna.utils.content_parser import ContentParser
-from apps.users.models import User
+
+# ==============================================================================
+# QUESTION LIST (GET /api/v1/qna/questions)
+# ==============================================================================
 
 
 class QuestionQuerySerializer(QnAVersionedValidationMixin, serializers.Serializer[Any]):
@@ -19,49 +26,6 @@ class QuestionQuerySerializer(QnAVersionedValidationMixin, serializers.Serialize
     sort = serializers.ChoiceField(choices=["latest", "oldest", "most_views"], default="latest")
     page = serializers.IntegerField(default=1)
     size = serializers.IntegerField(default=10)
-
-
-class QuestionCategoryListSerializer(serializers.ModelSerializer[QuestionCategory]):
-    """
-    질의응답 목록용 카테고리 시리얼라이저 (대 > 중 > 소 형태)
-    """
-
-    depth = serializers.SerializerMethodField()
-    names = serializers.SerializerMethodField()
-
-    class Meta:
-        model = QuestionCategory
-        fields = ["id", "depth", "names"]
-
-    def get_depth(self, obj: QuestionCategory) -> int:
-        """모델에 필드가 없으므로 부모를 거슬러 올라가며 깊이를 계산"""
-        depth = 0
-        curr = obj.parent
-        while curr:
-            depth += 1
-            curr = curr.parent
-        return depth
-
-    def get_names(self, obj: QuestionCategory) -> list[str]:
-        """부모 카테고리를 거슬러 올라가며 전체 경로 이름을 리스트로 생성"""
-        names: list[str] = []
-        curr: QuestionCategory | None = obj
-        while curr:
-            names.insert(0, curr.name)
-            curr = curr.parent
-        return names
-
-
-class QuestionAuthorSerializer(serializers.ModelSerializer[User]):
-    """
-    질문 작성자 정보 시리얼라이저
-    """
-
-    profile_image_url = serializers.ImageField(source="profile_img_url", use_url=True)
-
-    class Meta:
-        model = User
-        fields = ["id", "nickname", "profile_image_url"]
 
 
 class QuestionListSerializer(serializers.ModelSerializer[Question]):
@@ -98,16 +62,9 @@ class QuestionListSerializer(serializers.ModelSerializer[Question]):
         return ContentParser.extract_thumbnail_img_url(obj.content)
 
 
-class QuestionCreateSerializer(QnAVersionedValidationMixin, serializers.ModelSerializer[Question]):
-    """
-    질문 등록 시리얼라이저
-    """
-
-    category_id = serializers.IntegerField(required=True, help_text="카테고리 ID (소분류)")
-
-    class Meta:
-        model = Question
-        fields = ["title", "content", "category_id"]
+# ==============================================================================
+# ACTION RESPONSES (POST/PUT/DELETE Success)
+# ==============================================================================
 
 
 class QuestionCreateResponseSerializer(QnAVersionedValidationMixin, serializers.Serializer[Any]):
